requirements.txt
kafka-python 
pyyaml
